Split declared env effect tests
