+++ tests/integration/cli_build/declared_env_effects.rs
@@ -2,340 +2,14 @@
 mod build_graph;
 #[path = "declared_env_effects/emit.rs"]
 mod emit;
+#[path = "declared_env_effects/executable.rs"]
+mod executable;
 #[path = "declared_env_effects/rejections.rs"]
 mod rejections;
+#[path = "declared_env_effects/test_command.rs"]
+mod test_command;
 
 use std::process::Command;
-
-#[test]
-fn build_command_build_zen_accepts_declared_env_read_with_fallback() {
-    assert_executable_command_accepts_declared_env_read(
-        &["build", "build.zen"],
-        r#"| .Err { "default" }"#,
-        "build_command_build_zen_accepts_declared_env_read_with_fallback",
-        ExecutableCommandExpectation::BuildOutput,
-    );
-}
-
-#[test]
-fn build_command_build_zen_accepts_wildcard_fallback_declared_env_read() {
-    assert_executable_command_accepts_declared_env_read(
-        &["build", "build.zen"],
-        r#"| _ { "default" }"#,
-        "build_command_build_zen_accepts_wildcard_fallback_declared_env_read",
-        ExecutableCommandExpectation::BuildOutput,
-    );
-}
-
-#[test]
-fn build_command_build_zen_accepts_identifier_fallback_declared_env_read() {
-    assert_executable_command_accepts_declared_env_read(
-        &["build", "build.zen"],
-        r#"| err { "default" }"#,
-        "build_command_build_zen_accepts_identifier_fallback_declared_env_read",
-        ExecutableCommandExpectation::BuildOutput,
-    );
-}
-
-#[test]
-fn build_command_build_zen_accepts_declared_env_read_for_multiple_targets() {
-    assert_executable_command_accepts_declared_env_read_for_multiple_targets(
-        &["build", "build.zen"],
-        r#"| .Err { "default" }"#,
-        "build_command_build_zen_accepts_declared_env_read_for_multiple_targets",
-    );
-}
-
-#[test]
-fn build_command_build_zen_accepts_wildcard_fallback_declared_env_read_for_multiple_targets() {
-    assert_executable_command_accepts_declared_env_read_for_multiple_targets(
-        &["build", "build.zen"],
-        r#"| _ { "default" }"#,
-        "build_command_build_zen_accepts_wildcard_fallback_declared_env_read_for_multiple_targets",
-    );
-}
-
-#[test]
-fn build_command_build_zen_accepts_identifier_fallback_declared_env_read_for_multiple_targets() {
-    assert_executable_command_accepts_declared_env_read_for_multiple_targets(
-        &["build", "build.zen"],
-        r#"| err { "default" }"#,
-        "build_command_build_zen_accepts_identifier_fallback_declared_env_read_for_multiple_targets",
-    );
-}
-
-fn assert_executable_command_accepts_declared_env_read_for_multiple_targets(
-    args: &[&str],
-    fallback_arm: &str,
-    case_name: &str,
-) {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        format!(
-            r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {{
-    std_path = b.os.env("ZEN_STD") ?
-        | .Ok(value) {{ value }}
-        {fallback_arm}
-    b.add(Executable {{ name: "app", main: "app.zen", out_dir: "build/app/" }})
-    b.add(Executable {{ name: "tool", main: "tool.zen", out_dir: "build/tool/" }})
-    .Ok(b.config())
-}}
-"#,
-        ),
-    )
-    .expect("write build.zen");
-    std::fs::write(
-        tmp.path().join("app.zen"),
-        r#"
-main = () i32 {
-    0
-}
-"#,
-    )
-    .expect("write app.zen");
-    std::fs::write(
-        tmp.path().join("tool.zen"),
-        r#"
-main = () i32 {
-    0
-}
-"#,
-    )
-    .expect("write tool.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(args)
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen executable build graph command");
-
-    assert!(
-        output.status.success(),
-        "{case_name}: zen executable build graph command failed: stdout={}, stderr={}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    for bin_path in [
-        tmp.path().join("build").join("app").join("app"),
-        tmp.path().join("build").join("tool").join("tool"),
-    ] {
-        assert!(
-            bin_path.exists(),
-            "expected {} to exist",
-            bin_path.display()
-        );
-    }
-}
-
-#[test]
-fn direct_file_command_build_zen_accepts_declared_env_read_with_fallback() {
-    assert_executable_command_accepts_declared_env_read(
-        &["build.zen"],
-        r#"| .Err { "default" }"#,
-        "direct_file_command_build_zen_accepts_declared_env_read_with_fallback",
-        ExecutableCommandExpectation::BuildOutput,
-    );
-}
-
-#[test]
-fn direct_file_command_build_zen_accepts_wildcard_fallback_declared_env_read() {
-    assert_executable_command_accepts_declared_env_read(
-        &["build.zen"],
-        r#"| _ { "default" }"#,
-        "direct_file_command_build_zen_accepts_wildcard_fallback_declared_env_read",
-        ExecutableCommandExpectation::BuildOutput,
-    );
-}
-
-#[test]
-fn direct_file_command_build_zen_accepts_identifier_fallback_declared_env_read() {
-    assert_executable_command_accepts_declared_env_read(
-        &["build.zen"],
-        r#"| err { "default" }"#,
-        "direct_file_command_build_zen_accepts_identifier_fallback_declared_env_read",
-        ExecutableCommandExpectation::BuildOutput,
-    );
-}
-
-#[test]
-fn direct_file_command_build_zen_accepts_declared_env_read_for_multiple_targets() {
-    assert_executable_command_accepts_declared_env_read_for_multiple_targets(
-        &["build.zen"],
-        r#"| .Err { "default" }"#,
-        "direct_file_command_build_zen_accepts_declared_env_read_for_multiple_targets",
-    );
-}
-
-#[test]
-fn direct_file_command_build_zen_accepts_wildcard_fallback_declared_env_read_for_multiple_targets()
-{
-    assert_executable_command_accepts_declared_env_read_for_multiple_targets(
-        &["build.zen"],
-        r#"| _ { "default" }"#,
-        "direct_file_command_build_zen_accepts_wildcard_fallback_declared_env_read_for_multiple_targets",
-    );
-}
-
-#[test]
-fn direct_file_command_build_zen_accepts_identifier_fallback_declared_env_read_for_multiple_targets(
-) {
-    assert_executable_command_accepts_declared_env_read_for_multiple_targets(
-        &["build.zen"],
-        r#"| err { "default" }"#,
-        "direct_file_command_build_zen_accepts_identifier_fallback_declared_env_read_for_multiple_targets",
-    );
-}
-
-#[test]
-fn test_command_build_zen_accepts_declared_env_read_with_fallback() {
-    assert_test_command_accepts_declared_env_read(
-        r#"| .Err { "default" }"#,
-        "test_command_build_zen_accepts_declared_env_read_with_fallback",
-    );
-}
-
-#[test]
-fn test_command_build_zen_accepts_wildcard_fallback_declared_env_read() {
-    assert_test_command_accepts_declared_env_read(
-        r#"| _ { "default" }"#,
-        "test_command_build_zen_accepts_wildcard_fallback_declared_env_read",
-    );
-}
-
-#[test]
-fn test_command_build_zen_accepts_identifier_fallback_declared_env_read() {
-    assert_test_command_accepts_declared_env_read(
-        r#"| err { "default" }"#,
-        "test_command_build_zen_accepts_identifier_fallback_declared_env_read",
-    );
-}
-
-#[test]
-fn test_command_build_zen_accepts_declared_env_read_for_multiple_targets() {
-    assert_test_command_accepts_declared_env_read_for_multiple_targets(
-        r#"| .Err { "default" }"#,
-        "test_command_build_zen_accepts_declared_env_read_for_multiple_targets",
-    );
-}
-
-#[test]
-fn test_command_build_zen_accepts_wildcard_fallback_declared_env_read_for_multiple_targets() {
-    assert_test_command_accepts_declared_env_read_for_multiple_targets(
-        r#"| _ { "default" }"#,
-        "test_command_build_zen_accepts_wildcard_fallback_declared_env_read_for_multiple_targets",
-    );
-}
-
-#[test]
-fn test_command_build_zen_accepts_identifier_fallback_declared_env_read_for_multiple_targets() {
-    assert_test_command_accepts_declared_env_read_for_multiple_targets(
-        r#"| err { "default" }"#,
-        "test_command_build_zen_accepts_identifier_fallback_declared_env_read_for_multiple_targets",
-    );
-}
-
-fn assert_test_command_accepts_declared_env_read_for_multiple_targets(
-    fallback_arm: &str,
-    case_name: &str,
-) {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        format!(
-            r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {{
-    std_path = b.os.env("ZEN_STD") ?
-        | .Ok(value) {{ value }}
-        {fallback_arm}
-    b.add(Test {{ name: "unit", root: "unit.zen" }})
-    b.add(Test {{ name: "integration", root: "integration.zen" }})
-    .Ok(b.config())
-}}
-"#,
-        ),
-    )
-    .expect("write build.zen");
-    std::fs::write(
-        tmp.path().join("unit.zen"),
-        r#"
-main = () i32 {
-    0
-}
-"#,
-    )
-    .expect("write unit.zen");
-    std::fs::write(
-        tmp.path().join("integration.zen"),
-        r#"
-main = () i32 {
-    0
-}
-"#,
-    )
-    .expect("write integration.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(["test", "build.zen"])
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen test build.zen");
-
-    assert!(
-        output.status.success(),
-        "{case_name}: zen test build.zen failed: stdout={}, stderr={}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    assert!(
-        stdout.contains("test unit passed") && stdout.contains("test integration passed"),
-        "expected both test targets to pass, stdout={stdout}"
-    );
-}
-
-fn assert_test_command_accepts_declared_env_read(fallback_arm: &str, case_name: &str) {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        format!(
-            r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {{
-    std_path = b.os.env("ZEN_STD") ?
-        | .Ok(value) {{ value }}
-        {fallback_arm}
-    b.add(Test {{ name: "unit", root: "test.zen" }})
-    .Ok(b.config())
-}}
-"#,
-        ),
-    )
-    .expect("write build.zen");
-    std::fs::write(
-        tmp.path().join("test.zen"),
-        r#"
-main = () i32 {
-    0
-}
-"#,
-    )
-    .expect("write test.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(["test", "build.zen"])
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen test build.zen");
-
-    assert!(
-        output.status.success(),
-        "{case_name}: zen test build.zen failed: stdout={}, stderr={}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-}
 
 enum ExecutableCommandExpectation {
     BuildOutput,
